requirements.txt
curlify==2.2.1
hyundai_kia_connect_api==3.34.1
coloredlogs==15.0.1
flask==3.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pymysql==1.1.1
APScheduler==3.10.4